backend/src/main/java/INFSUS/project/PRO/controllers/LoginController.java
package INFSUS.project.PRO.controllers;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import INFSUS.project.PRO.models.User;
import INFSUS.project.PRO.repository.UserRepository;

@RestController
@CrossOrigin(origins = "http://localhost:3000", allowedHeaders = "*")
@RequestMapping("/login")
public class LoginController {

    @Autowired
    UserRepository userRepository;

    @GetMapping("")
    public ResponseEntity<?> login(@RequestParam String username) {
        User user = userRepository.findByUsername(username).orElse(null);

        if (user != null) {
            return ResponseEntity.ok(user.getUsername());
        } else {
            return ResponseEntity.badRequest().body("User not found.");
        }
    }

}
